+++ src/main/java/org/opensearch/knn/quantization/models/quantizationState/OneBitScalarQuantizationState.java
@@ -13,6 +13,7 @@ import org.opensearch.Version;
 import org.opensearch.core.common.io.stream.StreamInput;
 import org.opensearch.core.common.io.stream.StreamOutput;
 import org.opensearch.knn.quantization.models.quantizationParams.ScalarQuantizationParams;
+import static org.opensearch.knn.common.KNNConstants.BYTE_ALIGNMENT_MASK;
 
 import java.io.IOException;
 
@@ -117,14 +118,14 @@ public final class OneBitScalarQuantizationState implements QuantizationState {
     @Override
     public int getBytesPerVector() {
         // Calculate the number of bytes required for one-bit quantization
-        return meanThresholds.length;
+        return (meanThresholds.length + BYTE_ALIGNMENT_MASK) / Byte.SIZE;
     }
 
     @Override
     public int getDimensions() {
         // For one-bit quantization, the dimension for indexing is just the length of the thresholds array.
         // Align the original dimensions to the next multiple of 8
-        return (meanThresholds.length + 7) & ~7;
+        return (meanThresholds.length + BYTE_ALIGNMENT_MASK) & ~BYTE_ALIGNMENT_MASK;
     }
 
     /**